refactor: extract SsrfValidator to eliminate 4x duplicated SSRF logic
