+++ apps/api/Codec.Api/Program.cs
@@ -363,54 +363,7 @@ builder.Services.AddHttpClient<DiscordMediaRehostService>()
         MaxAutomaticRedirections = 3,
         UseCookies = false,
         PooledConnectionLifetime = TimeSpan.FromMinutes(5),
-        ConnectCallback = async (context, cancellationToken) =>
-        {
-            // DNS rebinding protection: resolve and validate before connecting.
-            var entries = await Dns.GetHostAddressesAsync(context.DnsEndPoint.Host, cancellationToken);
-            foreach (var entry in entries)
-            {
-                // Normalize IPv4-mapped IPv6 addresses (e.g. ::ffff:10.0.0.1) to IPv4
-                var ip = entry.IsIPv4MappedToIPv6 ? entry.MapToIPv4() : entry;
-
-                if (IPAddress.IsLoopback(ip) || ip.IsIPv6LinkLocal || ip.IsIPv6SiteLocal)
-                    throw new HttpRequestException($"Blocked rehost connection to private IP {ip}.");
-
-                if (ip.AddressFamily == AddressFamily.InterNetwork)
-                {
-                    var bytes = ip.GetAddressBytes();
-                    var isPrivate = bytes[0] switch
-                    {
-                        10 => true,
-                        172 => bytes[1] >= 16 && bytes[1] <= 31,
-                        192 => bytes[1] == 168,
-                        169 => bytes[1] == 254,
-                        127 => true,
-                        0 => true,
-                        _ => false
-                    };
-                    if (isPrivate)
-                        throw new HttpRequestException($"Blocked rehost connection to private IP {ip}.");
-                }
-                else if (ip.AddressFamily == AddressFamily.InterNetworkV6)
-                {
-                    var bytes = ip.GetAddressBytes();
-                    if (bytes[0] is 0xFC or 0xFD)
-                        throw new HttpRequestException($"Blocked rehost connection to private IP {ip}.");
-                }
-            }
-
-            var socket = new Socket(SocketType.Stream, ProtocolType.Tcp) { NoDelay = true };
-            try
-            {
-                await socket.ConnectAsync(new IPEndPoint(entries[0], context.DnsEndPoint.Port), cancellationToken);
-                return new NetworkStream(socket, ownsSocket: true);
-            }
-            catch
-            {
-                socket.Dispose();
-                throw;
-            }
-        }
+        ConnectCallback = SsrfValidator.CreateSafeConnectCallback("rehost")
     });
 
 // Web Push notification service (VAPID-authenticated).
@@ -446,53 +399,7 @@ builder.Services.AddHttpClient("webhook", client =>
     AllowAutoRedirect = false,
     UseCookies = false,
     PooledConnectionLifetime = TimeSpan.FromMinutes(5),
-    ConnectCallback = async (context, cancellationToken) =>
-    {
-        // DNS rebinding protection: resolve and validate before connecting.
-        var entries = await Dns.GetHostAddressesAsync(context.DnsEndPoint.Host, cancellationToken);
-        foreach (var entry in entries)
-        {
-            var ip = entry.IsIPv4MappedToIPv6 ? entry.MapToIPv4() : entry;
-
-            if (IPAddress.IsLoopback(ip) || ip.IsIPv6LinkLocal || ip.IsIPv6SiteLocal)
-                throw new HttpRequestException($"Blocked webhook connection to private IP {ip}.");
-
-            if (ip.AddressFamily == AddressFamily.InterNetwork)
-            {
-                var bytes = ip.GetAddressBytes();
-                var isPrivate = bytes[0] switch
-                {
-                    10 => true,
-                    172 => bytes[1] >= 16 && bytes[1] <= 31,
-                    192 => bytes[1] == 168,
-                    169 => bytes[1] == 254,
-                    127 => true,
-                    0 => true,
-                    _ => false
-                };
-                if (isPrivate)
-                    throw new HttpRequestException($"Blocked webhook connection to private IP {ip}.");
-            }
-            else if (ip.AddressFamily == AddressFamily.InterNetworkV6)
-            {
-                var bytes = ip.GetAddressBytes();
-                if (bytes[0] is 0xFC or 0xFD)
-                    throw new HttpRequestException($"Blocked webhook connection to private IP {ip}.");
-            }
-        }
-
-        var socket = new Socket(SocketType.Stream, ProtocolType.Tcp) { NoDelay = true };
-        try
-        {
-            await socket.ConnectAsync(new IPEndPoint(entries[0], context.DnsEndPoint.Port), cancellationToken);
-            return new NetworkStream(socket, ownsSocket: true);
-        }
-        catch
-        {
-            socket.Dispose();
-            throw;
-        }
-    }
+    ConnectCallback = SsrfValidator.CreateSafeConnectCallback("webhook")
 });
 
 // Link preview HttpClient with DNS rebinding protection, redirect limits, and no cookies.
@@ -507,60 +414,7 @@ builder.Services.AddHttpClient<ILinkPreviewService, LinkPreviewService>(client =
     MaxAutomaticRedirections = 3,
     UseCookies = false,
     PooledConnectionLifetime = TimeSpan.FromMinutes(5),
-    ConnectCallback = async (context, cancellationToken) =>
-    {
-        // DNS rebinding protection: resolve and validate before connecting.
-        var entries = await Dns.GetHostAddressesAsync(context.DnsEndPoint.Host, cancellationToken);
-        foreach (var entry in entries)
-        {
-            var ip = entry.IsIPv4MappedToIPv6 ? entry.MapToIPv4() : entry;
-
-            if (IPAddress.IsLoopback(ip) || ip.IsIPv6LinkLocal || ip.IsIPv6SiteLocal)
-            {
-                throw new HttpRequestException($"Blocked connection to private IP {ip}.");
-            }
-
-            if (ip.AddressFamily == AddressFamily.InterNetwork)
-            {
-                var bytes = ip.GetAddressBytes();
-                var isPrivate = bytes[0] switch
-                {
-                    10 => true,
-                    172 => bytes[1] >= 16 && bytes[1] <= 31,
-                    192 => bytes[1] == 168,
-                    169 => bytes[1] == 254,
-                    127 => true,
-                    0 => true,
-                    _ => false
-                };
-
-                if (isPrivate)
-                {
-                    throw new HttpRequestException($"Blocked connection to private IP {ip}.");
-                }
-            }
-            else if (ip.AddressFamily == AddressFamily.InterNetworkV6)
-            {
-                var bytes = ip.GetAddressBytes();
-                if (bytes[0] is 0xFC or 0xFD)
-                {
-                    throw new HttpRequestException($"Blocked connection to private IP {ip}.");
-                }
-            }
-        }
-
-        var socket = new Socket(SocketType.Stream, ProtocolType.Tcp) { NoDelay = true };
-        try
-        {
-            await socket.ConnectAsync(new IPEndPoint(entries[0], context.DnsEndPoint.Port), cancellationToken);
-            return new NetworkStream(socket, ownsSocket: true);
-        }
-        catch
-        {
-            socket.Dispose();
-            throw;
-        }
-    }
+    ConnectCallback = SsrfValidator.CreateSafeConnectCallback("link-preview")
 });
 
 // Image proxy HttpClient with DNS rebinding protection, redirect limits, and no cookies.
@@ -575,59 +429,7 @@ builder.Services.AddHttpClient<IImageProxyService, ImageProxyService>(client =>
     MaxAutomaticRedirections = 3,
     UseCookies = false,
     PooledConnectionLifetime = TimeSpan.FromMinutes(5),
-    ConnectCallback = async (context, cancellationToken) =>
-    {
-        var entries = await Dns.GetHostAddressesAsync(context.DnsEndPoint.Host, cancellationToken);
-        foreach (var entry in entries)
-        {
-            var ip = entry.IsIPv4MappedToIPv6 ? entry.MapToIPv4() : entry;
-
-            if (IPAddress.IsLoopback(ip) || ip.IsIPv6LinkLocal || ip.IsIPv6SiteLocal)
-            {
-                throw new HttpRequestException($"Blocked connection to private IP {ip}.");
-            }
-
-            if (ip.AddressFamily == AddressFamily.InterNetwork)
-            {
-                var bytes = ip.GetAddressBytes();
-                var isPrivate = bytes[0] switch
-                {
-                    10 => true,
-                    172 => bytes[1] >= 16 && bytes[1] <= 31,
-                    192 => bytes[1] == 168,
-                    169 => bytes[1] == 254,
-                    127 => true,
-                    0 => true,
-                    _ => false
-                };
-
-                if (isPrivate)
-                {
-                    throw new HttpRequestException($"Blocked connection to private IP {ip}.");
-                }
-            }
-            else if (ip.AddressFamily == AddressFamily.InterNetworkV6)
-            {
-                var bytes = ip.GetAddressBytes();
-                if (bytes[0] is 0xFC or 0xFD)
-                {
-                    throw new HttpRequestException($"Blocked connection to private IP {ip}.");
-                }
-            }
-        }
-
-        var socket = new Socket(SocketType.Stream, ProtocolType.Tcp) { NoDelay = true };
-        try
-        {
-            await socket.ConnectAsync(new IPEndPoint(entries[0], context.DnsEndPoint.Port), cancellationToken);
-            return new NetworkStream(socket, ownsSocket: true);
-        }
-        catch
-        {
-            socket.Dispose();
-            throw;
-        }
-    }
+    ConnectCallback = SsrfValidator.CreateSafeConnectCallback("image-proxy")
 });
 
 // GitHub Issues API client for in-app bug reports.
